introducing a caching parser and a factory
