+++ tla-types/src/test/scala/at/forsyte/apalache/tla/typecheck/TestCachingType1Parser.scala
@@ -1,0 +1,14 @@
+package at.forsyte.apalache.tla.typecheck
+
+import org.junit.runner.RunWith
+import org.scalatest.FunSuite
+import org.scalatest.junit.JUnitRunner
+
+@RunWith(classOf[JUnitRunner])
+class TestCachingType1Parser  extends FunSuite {
+  test("Int") {
+    val parser = TypeParserFactory.cacheType1Parser()
+    val result = parser("Int")
+    assert(IntT1() == result)
+  }
+}
